Moved createWebDriver to root test package
