+++ src/test/java/test/WebDriverFactory.kt
@@ -1,4 +1,4 @@
-package test.webdriverfactory
+package test
 
 import org.openqa.selenium.WebDriver
 import org.openqa.selenium.chrome.ChromeDriver
